Added namespacing to enable compatability with other libraries with functions of the same name.
